% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues_calculation.R
\name{cmh_cal}
\alias{cmh_cal}
\title{cmh_cal is the function of the cmh test for single genes with stratas}
\usage{
cmh_cal(dis, cur_g, strata_vector, k = 1, alter_option = "greater",
  exact_option = TRUE)
}
\arguments{
\item{dis}{is a n-length numeric vector of indicators of phenotypes from n samples.}

\item{cur_g}{is a n-length numeric vector of indicators of mutations from n samples in a certain gene.}

\item{strata_vector}{is a n-length numeric vector of categories from the k kinds of strata, it is converted from the result of strat_score_cal_glm, for example, when we have 5 categories, and when strata = [[1,0,0,0,0];[0,1,0,0,0];[0,0,1,0,0];[0,0,0,0,0]], we have: strata_vector = (2,3,4,0).}

\item{k}{is the kinds of strata, in accordance with content of strata_vector}

\item{alter_option}{decides the test for p-values calculation. options=c("greater","less","two.sided"):
"greater"(default):apply one-side cmh test if mutation are enriched in cases.
"less":apply one-side cmh test if mutation are enriched in controls.
"two.sided": apply two-side cmh test}

\item{exact_option}{A logical indicating whether the Mantel-Haenszel test(FALSE) or the exact conditional test (TRUE, default) is applied.If k==1, exact_option will be ignored.}
}
\value{
a p-value from cmh test.
}
\description{
cmh_cal belongs to pval_ss_cal
when there is only 1 strata, cmh_cal reduces to fisher exact test and ignore the exact_option
}
\examples{
pheno=rbinom(100,1,0.5)
cur_pc=rbind(matrix(rnorm(500,0,1),ncol=10,nrow=50),matrix(rnorm(500,0.5,1),ncol=10,nrow=50))
strata=strat_score_cal_glm(pheno,cur_pc)
cur_geno=rbinom(100,1,0.1)
cur_pval=cmh_cal(pheno,cur_geno,strata, alter="greater",exact_option=TRUE)
}
\references{
William G. Cochran (December 1954). "Some Methods for Strengthening the Common chi-squared Tests". Biometrics. 10 (4): 417–451. doi:10.2307/3001616. JSTOR 3001616.

Nathan Mantel and William Haenszel (April 1959). "Statistical aspects of the analysis of data from retrospective studies of disease". Journal of the National Cancer Institute. 22 (4): 719–748. doi:10.1093/jnci/22.4.719. PMID 13655060.

Fisher, R. A. (1922). "On the interpretation of chi-squared from contingency tables, and the calculation of P". Journal of the Royal Statistical Society. 85 (1): 87–94. doi:10.2307/2340521. JSTOR 2340521.
}
\seealso{
\code{\link{mantelhaen.test}} which this function wraps
}
