% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdminp_calculation.R
\name{find_rank}
\alias{find_rank}
\title{This function calculates the ranks of a vector based on another vector}
\usage{
find_rank(target, ruler)
}
\arguments{
\item{target}{a decreacing-sorted vector for measure by the ruler}

\item{ruler}{has the same length as target, also a decreasing-sorted vector}
}
\value{
a numerica vector, with each element i shows the numbers of values in the target that is bigger than the ruler[i]
}
\description{
find_rank is used by sdminp_fdr to speed up the calculation
}
\examples{
a=c(8,6,4,2)
b=c(7,4,3,1)
find_rank(a,b)
}
