% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calculation.R
\name{get_centrality}
\alias{get_centrality}
\title{get_centrality match the centrality prior information to the given gene set from collection of MSigDB}
\usage{
get_centrality(interact_m, w_option = "deg", direct_option = FALSE,
  mode_option = "all")
}
\arguments{
\item{interact_m}{represents the genetic network. It is the adjacency matrix of a graph with each node represents a gene.}

\item{w_option}{the kind of centralities.can be "deg" for degree,"closn" for closeness,"betn" for betweenness, "eigen" for eigervector centrality, and "pagerank"}

\item{direct_option}{if it is true, the network will be calculated as directed pathways, parameter especially for pagerank}

\item{mode_option}{parameters for centrality calculation, "out" for out-degree, "in" for in-degree or "all" or "total" for the sum of the two. see \code{\link{igraph}} for more details.}
}
\description{
Different from match_prior_info, the match_prior_info_centrality calculates each interactions based on the true gene sets.
}
\references{
Page, L., Brin, S., Motwani, R., & Winograd, T. (1999). The PageRank citation ranking: Bringing order to the web. Stanford InfoLab. Retrieved from http://ilpubs.stanford.edu:8090/422

White, S., & Smyth, P. (2003). Algorithms for Estimating Relative Importance in Networks. In Proceedings of the Ninth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (pp. 266–275). New York, NY, USA: ACM. https://doi.org/10.1145/956750.956782
}
