% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calculation.R
\name{get_gene_expression}
\alias{get_gene_expression}
\title{This function get gene expression data of specific tissues from GTEx}
\usage{
get_gene_expression(
  gene_label = "ensembl_gene_id",
  tissue = support_gtex_tissue(),
  comb = "none"
)
}
\arguments{
\item{gene_label}{is the gene names of the returning vector or matrix.It can be "ensembl_gene_id" or "symbols"}

\item{tissue}{a vector of filters the expression level of specific tissues. Use \code{\link{support_gtex_tissue()}} to see supported tissues.}

\item{comb}{is the operation on combining selected categories of tissues.
It can be "median","mean","max","min",and "none", which calculate the median, mean of genes or do nothing on them.}
}
\value{
the numeric vector or matrix(@param mode is "none") representing prior information for each single gene
}
\description{
The data resources comes from tpm of genes counts of RNAseq data of GTEx (gtexportal.org/home/datasets).
It based on GTEx_Analysis_2016-01-15_v7_RNASeQCv1.1.8_gene_tpm.gct.gz (Retrive Nov (2013))
}
\examples{
prior_expression=get_gene_expression(gene_label="symbols",tissue=c("Liver","Lung"),comb="mean")
prior_expression=get_gene_expression()
}
\references{
Lonsdale, John, et al. "The genotype-tissue expression (GTEx) project." Nature genetics 45.6 (2013): 580.
}
