% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calculation.R
\name{get_gene_length}
\alias{get_gene_length}
\title{This function estimates the transcripts length.}
\usage{
get_gene_length(gene_label = "symbols", comb = "mean")
}
\arguments{
\item{gene_label}{is the returning gene labels, it can be "ensembl_gene_id" or "symbols"}

\item{comb}{is the estimation method based on multiple records of transcription lengths.
It can be "median","mean","max","min", which calculate the median, mean,max and min of genes or do nothing on them.}
}
\value{
the numeric vector representing prior information for each single gene's length
}
\description{
The data resources comes from the ensembl genome browser (useast.ensembl.org/index.html).
}
\examples{
prior_length=get_gene_length(gene_label="symbols",comb="median")
}
\references{
Durinck, Steffen, et al. "Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt." Nature protocols 4.8 (2009): 1184.
}
\seealso{
\code{\link{biomaRt}}
}
