% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calculation.R
\name{get_genic_intolerance}
\alias{get_genic_intolerance}
\title{This function pull down the genic intolerance information}
\usage{
get_genic_intolerance()
}
\value{
the numeric vector representing prior information for each single gene, with genes name corresponding to that in the net, which can be used as input of function match_prior_info
}
\description{
The data resources comes from the databased of genic intolerance.
(genic-intolerance.org/data/RVIS_Unpublished_ExAC_May2015.txt).
The default gene symbol is "CCDS_r15" and "%RVIS_ExAc_0.05%(AnyPopn)"(the percentage of genic intolerance)
The genes are labeled with gene symbols.
}
\examples{
genic_intolerance=get_genic_intolerance()

}
\references{
Petrovski, S., Wang, Q., Heinzen, E. L., Allen, A. S., & Goldstein, D. B. (2013). Genic Intolerance to Functional Variation and the Interpretation of Personal Genomes. PLOS Genetics, 9(8), e1003709. https://doi.org/10.1371/journal.pgen.1003709
}
