% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calculation.R
\name{match_prior_info}
\alias{match_prior_info}
\title{match_prior_info match the prior information to the given gene set from collection of MSigDB}
\usage{
match_prior_info(net, prior_info, add_option = "none",
  report_option = TRUE)
}
\arguments{
\item{net}{dataframe the gmt file from collections of Molecular signatures database (MSigDB), broad institute.}

\item{prior_info}{the numeric vector representing prior information for each single gene, with genes name corresponding to that in the net.}

\item{add_option}{defines the method of adding up missing values. It can be "none", "mean" or "median". No actions for adding up missing values if "none".}

\item{report_option}{if TRUE, report current procedures of the path, which is the proportion of sets completed the matching steps.}
}
\value{
a dataframe with the same format as net,which is the gmt files
}
\description{
match_prior_info match the prior information to the given gene set from collection of MSigDB
}
\examples{
net=net.h.all.v6.1.symbols
#net from MSigDB (software.broadinstitute.org/gsea/msigdb): "h.all.v6.1.symbols.gmt"
prior_gi=get_genic_intolerance()
prior_net=match_prior_info(net,prior_gi)
}
\references{
Liberzon, A., Subramanian, A., Pinchback, R., Thorvaldsdóttir, H., Tamayo, P., & Mesirov, J. P. (2011). Molecular signatures database (MSigDB) 3.0. Bioinformatics, 27(12), 1739–1740. https://doi.org/10.1093/bioinformatics/btr260
}
