% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calculation.R
\name{match_prior_info_centrality}
\alias{match_prior_info_centrality}
\title{match_prior_info_centrality matches the centrality prior information to the given gene set from collection of MSigDB}
\usage{
match_prior_info_centrality(
  net,
  human_whole,
  add_option = "none",
  report_option = TRUE,
  w_option = "deg",
  direct_option = FALSE,
  mode_option = "all"
)
}
\arguments{
\item{net}{dataframe the gmt file from collections of MSigDB, broad institute. each line represents a pathway.please read in with read.csv with header=FALSE and stringAsFactors = FALSE.}

\item{human_whole}{the 2-column matrix with each line representing the connection from gene in column 1 to gene in column 2}

\item{add_option}{defines the method of adding up missing values. It can be "none", "mean" or "median". No actions for adding up missing values if "none".}

\item{report_option}{if TRUE, report current procedures of the path, which is the proportion of sets completed the matching steps.}

\item{w_option}{the kind of centralities.}

\item{direct_option}{if it is true, the network will be calculated as directed pathways, parameter especially for pagerank}

\item{mode_option}{parameters for centrality calculation, "out" for out-degree, "in" for in-degree or "all" or "total" for the sum of the two.}
}
\value{
a dataframe with the same format as net,which is the gmt files
}
\description{
Different from match_prior_info, the match_prior_info_centrality calculates each interactions based on the true gene sets.
}
\examples{
net=net.h.all.v6.1.entrez
#from MSigDB (http://software.broadinstitute.org/gsea/msigdb): "h.all.v6.1.entrez.gmt"
human_whole=human_whole_biogird_3.4.147
#from BioGRID(https://thebiogrid.org/): "BIOGRID-ORGANISM-Homo_sapiens-3.4.147.tab2.txt"
human_whole=as.matrix(human_whole[,c(2,3,8,9,10,11)])
human_whole=unique(human_whole)
human_whole=as.matrix(human_whole[order(human_whole[,2]),])
human_whole=as.matrix(human_whole[order(human_whole[,1]),])
human_whole[,1]=as.numeric(human_whole[,1])
human_whole[,2]=as.numeric(human_whole[,2])
human_whole=human_whole[,1:2]
res=match_prior_info_centrality(net,human_whole,add_option="none",
report_option=TRUE,w_option="pagerank",direct_option=TRUE,mode_option="all")
}
\references{
Liberzon, A., Subramanian, A., Pinchback, R., Thorvaldsdóttir, H., Tamayo, P., & Mesirov, J. P. (2011). Molecular signatures database (MSigDB) 3.0. Bioinformatics, 27(12), 1739–1740. https://doi.org/10.1093/bioinformatics/btr260

Page, L., Brin, S., Motwani, R., & Winograd, T. (1999). The PageRank citation ranking: Bringing order to the web. Stanford InfoLab. Retrieved from http://ilpubs.stanford.edu:8090/422

White, S., & Smyth, P. (2003). Algorithms for Estimating Relative Importance in Networks. In Proceedings of the Ninth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (pp. 266–275). New York, NY, USA: ACM. https://doi.org/10.1145/956750.956782

Chatr-aryamontri, A., Oughtred, R., Boucher, L., Rust, J., Chang, C., Kolas, N. K., … Tyers, M. (2017). The BioGRID interaction database: 2017 update. Nucleic Acids Research, 45(Database issue), D369–D379. https://doi.org/10.1093/nar/gkw1102
}
\seealso{
\code{\link{igraph}} which this function wraps
}
