% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdminp_calculation.R
\name{rawp_cal}
\alias{rawp_cal}
\title{This function calculates the impirical raw p-values.}
\usage{
rawp_cal(res_ob, res_perm)
}
\arguments{
\item{res_ob}{is a numeric vectors of statistics}

\item{res_perm}{is a matrix with each colum is a permuated statistics with the same length as res_ob}
}
\value{
a numeric matrix with raw pvalues, defined by Ge et al, 2003 (box 1). NAs will be ignored.
}
\description{
It implements the raw-pvalues based on permutation, defined by Ge et al, 2003 (box 1)
}
\examples{
ob=rnorm(4,2,2)
perm=matrix(rnorm(20,2,2),ncol=5,nrow=4)
rawp_cal(ob,perm)
}
\references{
Ge, Y., Dudoit, S., & Speed, T. P. (2003). Resampling-based multiple testing for microarray data analysis. Test, 12(1), 1–77. https://doi.org/10.1007/BF02595811
}
