% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdminp_calculation.R
\name{sdminp_fdr}
\alias{sdminp_fdr}
\title{This function calculates the Step Down minP and report the results,
This is downstream analysis for sdminp
It implements the improved step-down minP algorithm based on FDR by Ge et al, 2003 (box 5)
It requires another function find_rank}
\usage{
sdminp_fdr(res_ob, res_perm, tao0 = 0.2)
}
\arguments{
\item{res_ob}{is a numeric vectors of statistics}

\item{res_perm}{is a matrix with each colum is a permuated statistics with the same length as res_ob}

\item{tao0}{is a proportion threshold that more than tao0, there will expected to be no significant results.}
}
\value{
a numeric matrix with adjusted pvalues,
 the first column is the FDR adjusted pvalues,
 the second colum is the corresponding q-values
}
\description{
This function calculates the Step Down minP and report the results,
This is downstream analysis for sdminp
It implements the improved step-down minP algorithm based on FDR by Ge et al, 2003 (box 5)
It requires another function find_rank
}
\examples{
ob=rnorm(10,2,2)
perm=matrix(rnorm(100,2,2),ncol=10,nrow=10)
sdminp_fdr(ob,perm)
}
\references{
Ge, Y., Dudoit, S., & Speed, T. P. (2003). Resampling-based multiple testing for microarray data analysis. Test, 12(1), 1–77. https://doi.org/10.1007/BF02595811
}
