% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calculation.R
\name{support_gtex_tissue}
\alias{support_gtex_tissue}
\title{support_gtex_tissue provides supported option for "tissue" in function get_gene_expression}
\usage{
support_gtex_tissue()
}
\description{
The data resources comes from tpm of genes counts of RNAseq data of "https://gtexportal.org/home/datasets".
}
\examples{
support_gtex_tissue()

}
\references{
Lonsdale, John, et al. "The genotype-tissue expression (GTEx) project." Nature genetics 45.6 (2013): 580.

https://storage.googleapis.com/gtex_analysis_v7/rna_seq_data/GTEx_Analysis_2016-01-15_v7_RNASeQCv1.1.8_gene_tpm.gct.gz (Retrive Nov (2013)).
}
\seealso{
\code{\link{get_gene_expression}}
}
