% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_criticism.R
\name{trans_w}
\alias{trans_w}
\title{This function transfers prior information into the weights}
\usage{
trans_w(w)
}
\arguments{
\item{w}{a numeric vector which is the original statistic for weight,need to be non-negative}
}
\value{
a numeric vector of the transfered weight with around mean~1, min~0.05.
 Missing values in input are transfered into 1
}
\description{
The weight is calculated with the equation: w=1/(a x prior_info+b x mean(prior_info)) and then scaled into mean(w)=1. Here we take a=0.95 and b=0.05.
}
\examples{
a=c(NA,exp(rnorm(7,0,1)))
trans_w(a)
}
