% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{gof_cal}
\alias{gof_cal}
\title{This function calculates the goodness-of-fit tests}
\usage{
gof_cal(
  pm,
  t0ratio = 0.1,
  gof_method = NA,
  single_statistic = "score",
  accumulate_option = "max",
  filter = 0,
  weight_option = "none",
  weight = 1
)
}
\arguments{
\item{pm}{is a statistics matrix of P-values or weighted pvalues, each row represents a gene (independent tests) and each column represents a dataset (e.g. a permutation or an observation). Pm are not encouraged to have only 1 rows, if that happend, warning massage will produced.}

\item{t0ratio}{is the ratio for the region c(0,t0ratio) of pvalues for statistic calculation.}

\item{gof_method}{is the option for the set based analysis methods, available option includes: "higher_criticism","berk_jones","skat","einmahl_mckeague", see our publication for more details.}

\item{single_statistic}{is the option for single statistic, it can be c("score","likelihood_ratio","wald"),it would be ignored is gof_method is assigned}

\item{accumulate_option}{is the option for combining method, it can be c("max","sum","rsum","topsum","max2","sum2","rsum2","topsum2"), it would be ignored is gof_method is assigned, the "max2","sum2","rsum2","topsum2" will be replaced into "max","sum","rsum","topsum" if gof method is assigned.}

\item{filter}{is the threshold to exclude extremely small pvalues to avoid them driving all signals.default 0.}

\item{weight_option}{defines the external prior information as weight. It can be "none", "in" and "out". "none" assigns no weight, "in" assigns weight towards each single pvalues and "out" assigns weight towards each statistic T.}

\item{weight}{is a vector which provides weight towards each genes.}
}
\value{
a numeric vector with each elements is a Higher criticism values calculated from each colum of the Pm
}
\description{
For ordered p-values \eqn{p_{(1)}<p_{(2)}<...<p_{(n)}}.Define \eqn{ Sn(t)=\sum_{i=1}^n 1_{p_{(i)}<=t} } .
Then define the statistic T as
\eqn{T=(Sn(p_{(i)})-n x p_{(i)})/\sqrt(n x p_{(i)} x (1-p_{(i)}))} for score statistic. or
\eqn{T=(Sn(p_{(i)})-n x p_{(i)})/\sqrt{Sn(p_{(i)})*(n-Sn(p_{(i)}))/n}} for wald statistic. or
\eqn{T=n (p_{(i)}  log(p_{(i)} n / Sn(p_{(i)}))+(1-p_{(i)})log((1-p_{(i)})n/(n-Sn(p_{(i)}))))}  for log likelihood ratio
}
\details{
The max method calculates the \eqn{Tmax=max T__{(i)} where 0 < i < t0ration x n}
The sum method calculates the \eqn{Tsum=sum T__{(i)} where 0 < i < t0ration x n}
The rsum is calculated with \eqn{Trsum=1/n x sum T__{(i)}. where 0<i<=(i_max)},
where i_max is the index of the maxium of Ts.
The topsum is calculated with \eqn{Ttopsum=1/n max T_(r)}.
where r belongs to the subset that Ts in the subsets are the top t0ratio proportion among all Ts.
}
\examples{
a=matrix(runif(200,0,1),ncol=4,nrow=50)
gof_cal(a)
}
\references{
Donoho, D., & Jin, J. (2004). Higher Criticism for Detecting Sparse Heterogeneous Mixtures. The Annals of Statistics, 32(3), 962–994.
}
